Split Tinkerbell provider sources by purpose
